% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forstringr-package.R
\docType{package}
\name{forstringr-package}
\alias{forstringr}
\alias{forstringr-package}
\title{forstringr: String Manipulation Package for Those Familiar with 'Microsoft Excel'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'forstringr' is to enable complex string manipulation in R especially to those more familiar with LEFT(), RIGHT(), and MID() functions in Microsoft Excel. The package combines the power of 'stringr' with other manipulation packages such as 'dplyr' and 'tidyr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gbganalyst/forstringr}
  \item Report bugs at \url{https://github.com/gbganalyst/forstringr/issues}
}

}
\author{
\strong{Maintainer}: Ezekiel Ogundepo \email{gbganalyst@gmail.com} (\href{https://orcid.org/0000-0003-3974-2733}{ORCID})

Other contributors:
\itemize{
  \item Olubukunola Oyedele \email{olubukunola.oyedele@54gene.com} [contributor]
  \item Fatimo Adebanjo \email{adebanjofatimo2000@gmail.com} [contributor]
}

}
\keyword{internal}
