% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_omit_na.R
\name{length_omit_na}
\alias{length_omit_na}
\title{Length of an object}
\usage{
length_omit_na(x)
}
\arguments{
\item{x}{Input vector. Either a vector, or something coercible to one.}
}
\value{
An integer
}
\description{
\code{length_omitna()} counts only non-missing elements of a vector.
}
\examples{

ethnicity <- c("Hausa", NA, "Yoruba", "Igbo", NA, "Fulani", "Kanuri", "Others")

length_omit_na(ethnicity)

length(ethnicity)

}
\seealso{
\code{\link[=length]{length()}} counts all the elements in a vector including those that are missing (\link{NA}s).
}
