% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richest_in_nigeria.R
\docType{data}
\name{richest_in_nigeria}
\alias{richest_in_nigeria}
\title{Rank of billionaires in Nigeria}
\format{
A data frame with \code{10} rows and \code{5} variables:
\describe{
\item{Rank}{rank from 1 to 10}
\item{Name}{full name of the billionaires}
\item{Net worth}{net worth in billion dollars}
\item{Age}{the current age of billionaires}
\item{Source of Wealth}{the origin of the billionaires' entire body of wealth}
}
}
\source{
\url{https://rnn.ng/richest-men-in-nigeria/}
}
\usage{
richest_in_nigeria
}
\description{
A dataset containing the list of top ten billionaires in Nigeria.
}
\keyword{datasets}
