% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_left.R
\name{str_left}
\alias{str_left}
\title{Returns a substring from the beginning of a specified string}
\usage{
str_left(string, n = 1)
}
\arguments{
\item{string}{The character from which the left portion will be returned.}

\item{n}{Optional. The number of characters to return from the left side of string}
}
\value{
A character vector
}
\description{
Given a character vector, \code{str_left()} returns the left side of a string.
}
\examples{
str_left("Nigeria")
str_left("Nigeria", n = 3)
str_left(c("Female", "Male", "Male", "Female"))

}
\seealso{
\code{\link[=str_right]{str_right()}} which extracts characters from the right and \code{\link[=str_mid]{str_mid()}} which returns a segment of character strings.
}
