% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_mid.R
\name{str_mid}
\alias{str_mid}
\title{Returns a segment of character strings}
\usage{
str_mid(string, start, n)
}
\arguments{
\item{string}{The text string containing the characters you want to extract.}

\item{start}{The position of the first character you want to extract in the text. The first character in text has \code{start = 1}, and so on.}

\item{n}{The length of character to extract.}
}
\value{
A character vector.
}
\description{
\code{str_mid()} returns a specific number of characters from a text string, starting at the position you specify, based on the number of characters you specify.
}
\examples{
str_mid("Super Eagle", 7, 5)

str_mid("Oyo Ibadan", 5, 6)

}
\seealso{
\code{\link[=str_left]{str_left()}} which extracts characters from the left and \code{\link[=str_right]{str_right()}} which extracts characters from the right.
}
