% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_rm_whitespace.R
\name{str_rm_whitespace_df}
\alias{str_rm_whitespace_df}
\title{Remove extra spaces in a data frame}
\usage{
str_rm_whitespace_df(df)
}
\arguments{
\item{df}{A data frame or data frame extension (e.g. a tibble) with leading or trailing spaces.}
}
\value{
A clean data frame with no leading or trailing spaces.
}
\description{
\code{str_rm_whitespace_df()} removes all leading, trailing, and collapses multiple consecutive white spaces in non-numerical variables in a data frame.
}
\examples{
richest_in_nigeria

str_rm_whitespace_df(richest_in_nigeria)

}
